feat: separanting the lauer into layers [S] - Single Responsibily Principle
