+++ SOLIDPrincipiosNaPratica/Repositorio/Interface/IServicoUsuario.cs
@@ -1,0 +1,6 @@
+﻿namespace SOLIDPrincipiosNaPratica.Repositorio.Interface;
+
+public interface IServicoUsuario
+{
+    void RegistrarUsuario(string nomeUsuario, string email);
+}
