+++ SOLIDPrincipiosNaPratica/SOLIDPrincipiosNaPratica.sln
@@ -1,0 +1,24 @@
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.5.2.0
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SOLIDPrincipiosNaPratica", "SOLIDPrincipiosNaPratica.csproj", "{5E241682-809E-2E52-307F-2048358F50A4}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{5E241682-809E-2E52-307F-2048358F50A4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5E241682-809E-2E52-307F-2048358F50A4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5E241682-809E-2E52-307F-2048358F50A4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5E241682-809E-2E52-307F-2048358F50A4}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {1BD554E8-F7A9-4B41-8020-306B64509EF9}
+	EndGlobalSection
+EndGlobal
